+++ Beam/app/Providers/EventServiceProvider.php
@@ -2,20 +2,17 @@
 
 namespace App\Providers;
 
-use App\Observers\GenderBalanceObserver;
-use Illuminate\Support\Facades\Event;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
-use Remp\BeamModule\Model\Article;
 
 class EventServiceProvider extends ServiceProvider
 {
     /**
      * The event listener mappings for the application.
      *
-     * @var array
+     * @var array<string, array<int, string>>
      */
     protected $listen = [
-        'App\Events\SomeEvent' => [
+        'App\Events\Event' => [
             'App\Listeners\EventListener',
         ],
     ];
@@ -27,6 +24,5 @@ class EventServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        Article::observe(GenderBalanceObserver::class);
     }
 }