+++ Sso/app/Http/Middleware/VerifyCsrfToken.php
@@ -9,8 +9,8 @@ class VerifyCsrfToken extends Middleware
     /**
      * The URIs that should be excluded from CSRF verification.
      *
-     * @var array
-     */
+     * @var array<int, string>
+ */
     protected $except = [
         //
     ];